Chor: Rearreanged folder structure
